djangoreactproject/requirements.txt
asgiref==3.7.2
Django==3.2.13
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
